[qm] move the qm enrichers to a seperate module
This commit moves all the qm enrichers to a dedicated
module to improve maintainability.
