+++ grimoire_elk/utils.py
@@ -91,12 +91,12 @@ from .enriched.finosmeetings import FinosMeetingsEnrich
 from .enriched.functest import FunctestEnrich
 from .enriched.gerrit import GerritEnrich
 from .enriched.git import GitEnrich
-from .enriched.gitqm import GitQMEnrich
+from grimoire_elk.enriched.qm.gitqm import GitQMEnrich
 from .enriched.github import GitHubEnrich
 from .enriched.githubql import GitHubQLEnrich
 from .enriched.github2 import GitHubEnrich2
 from .enriched.gitlab import GitLabEnrich
-from .enriched.gitlabqm import GitLabQMEnrich
+from grimoire_elk.enriched.qm.gitlabqm import GitLabQMEnrich
 from .enriched.gitter import GitterEnrich
 from .enriched.google_hits import GoogleHitsEnrich
 from .enriched.groupsio import GroupsioEnrich
@@ -109,13 +109,13 @@ from .enriched.mattermost import MattermostEnrich
 from .enriched.mbox import MBoxEnrich
 from .enriched.mediawiki import MediaWikiEnrich
 from .enriched.meetup import MeetupEnrich
-from .enriched.meetupqm import MeetupQMEnrich
+from grimoire_elk.enriched.qm.meetupqm import MeetupQMEnrich
 from .enriched.mozillaclub import MozillaClubEnrich
 from .enriched.nntp import NNTPEnrich
 from .enriched.pagure import PagureEnrich
 from .enriched.phabricator import PhabricatorEnrich
 from .enriched.pipermail import PipermailEnrich
-from .enriched.pipermailqm import PipermailQMEnrich
+from grimoire_elk.enriched.qm.pipermailqm import PipermailQMEnrich
 from .enriched.puppetforge import PuppetForgeEnrich
 from .enriched.redmine import RedmineEnrich
 from .enriched.remo import ReMoEnrich
