+++ setup.py
@@ -56,7 +56,13 @@ setup(name="grimoire-elk",
           'Programming Language :: Python :: 3.4',
           'Programming Language :: Python :: 3.5'],
       keywords="development repositories analytics",
-      packages=['grimoire_elk', 'grimoire_elk.enriched', 'grimoire_elk.raw', 'grimoire_elk.identities'],
+      packages=[
+          'grimoire_elk',
+          'grimoire_elk.enriched',
+          'grimoire_elk.raw',
+          'grimoire_elk.identities',
+          'grimoire_elk.qm'
+      ],
       package_dir={'grimoire_elk.enriched': 'grimoire_elk/enriched'},
       package_data={'grimoire_elk.enriched': ['mappings/*.json']},
       python_requires='>=3.4',
